requirements.txt
absl-py==0.7.1
astor==0.7.1
boto3==1.9.130
botocore==1.12.130
certifi==2019.3.9
chardet==3.0.4
Django==2.2
djangorestframework==3.9.2
docutils==0.14
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
idna==2.8
jmespath==0.9.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1
mock==2.0.0
numpy==1.16.2
pbr==5.1.3
Pillow==6.0.0
protobuf==3.7.1
pyrouge==0.1.3
python-dateutil==2.8.0
pytorch-pretrained-bert==0.6.1
pytz==2018.9
regex==2019.3.12
requests==2.21.0
s3transfer==0.2.0
six==1.12.0
sqlparse==0.3.0
tensorboard==1.13.1
tensorboardX==1.6
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
torch==1.0.1.post2
torchvision==0.2.2.post3
tqdm==4.31.1
urllib3==1.24.1
Werkzeug==0.15.2
